Refactor Block.tsx to remove unecessary width computations
