+++ frontend/src/components/widgets/Form/styled-components.ts
@@ -17,24 +17,14 @@
 
 import styled from "@emotion/styled"
 
-export interface StyledFormProps {
-  width: number
-}
-
 export const StyledFormSubmitContent = styled.div(() => ({
   display: "flex",
 }))
 
-export const StyledForm = styled.div<StyledFormProps>(({ width, theme }) => ({
+export const StyledForm = styled.div(({ theme }) => ({
   border: `1px solid ${theme.colors.fadedText10}`,
   borderRadius: theme.radii.md,
   padding: "calc(1em - 1px)", // 1px to account for border.
-  width,
-  // Same margin as StyledElementContainer.
-  marginTop: 0,
-  marginRight: 0,
-  marginBottom: theme.spacing.lg,
-  marginLeft: 0,
 }))
 
 export const StyledErrorContainer = styled.div(({ theme }) => ({
